requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyMuPDF==1.23.9
python-multipart==0.0.6
Pillow==11.3.0
